History: Lower default historyCountMax to 10
As it turns out, even with the CSS changes, 15 still doesn't fit on high-res monitors with text/image scaling.

14 would fit but it's a weird number that'd raise questions.

It seems like there's no way to make it fit while still fitting in the 'default' frame size.

This also resets the variable based on presence of another new one. This can be removed after testing.