History: Update ant history formatting
This replaces the terrible out-of-place "Plus.png" I made with the
default multiplier icon, which doesn't fit completely, but should be
clear and obvious to anyone by that point.
Also replaces the word "ELO" with an ELO pic, ripped from research.

This makes the formatting look a bit more consistent with itself and the
rest of the game's UI.