Force number formatting to be in English
Ever since the BigInt perf change, users on non-English locales have
been getting their local locale for *only* the thousands separator.
Some of these languages use a dot for thousands, and the decimal
formatting code always uses a dot for decimals.
The result is that you can't tell if "1.001" is 1001 or 1.001.

Using localized formats for both the thousands & decimal would be ideal
but comes with a different problem: locales & languages don't agree
what "billion" means. In some locales it's 1e9, in others it's 1e12.
https://en.wikipedia.org/wiki/Long_and_short_scales#Current_usage

Formatting one part of the number differently from the other is likely
to lead to even more confusion than there already is, and fixing this
up ourselves would be incredibly hard.
Also, the fallback for Safari 13 forces en-US formatting anyway.

Ergo, the easiest fix is to just force it to en-US, like the rest of
the game is written to and how it used to be.

(Also uses Intl.NumberFormat which should be faster when "formatting
large numbers of numbers" according to MDN, no detectable impact)