When autobuying research with roomba, stop on failure
Calculating cost and updating the DOM is complicated and causes a
small performance impact when done 33 times each tick.

This particular impact is much more observable in Firefox than in Chrome,
which seems to optimize it better, but it's still wasteful.

This reduces it to just once a tick for the majority of the game's time.