Refactor talisman autobuyer for performance
Currently, the talisman enhance & fortify autobuyers are being called
on each game tick. These recalculate a lot of game state even if nothing
is actually bought, and the design makes it do it up to 14 times/tick.

This reworks the autobuyer to recalculate once and only if at least one
kind of upgrade went through this tick.