matching formula to that in the Achievement
the formula on Achievements.ts#L488 refers to a small percentage boost limited to 15%, while the existing formula calls min with only a single parameter then multiplies by 15%, and ends up being greater than 1. I'm just copying the formula from the description, and adding a /100 to set it as a percentage.