Only autobuy one tesseract building at a time
This fixes the bug mentioned in #68 (by not buying multiples at all),
and fixes some edge cases with the previous logic.

As a counterexample to the previous logic, imagine:
- No buildings bought (costs are [1, 10, 100, ...])
- tesseractbuyamount is 10.
- Player has 100 tesseracts.
The previous logic would buy two tier 2s (10+80=90), which is cheaper
than buying four tier 1s (1+8+27+64=100) or one tier 3 (100).
However, the expected behavior here is to buy three tier 1s and one tier
2.

As a result, however, autobuying is slower and takes N ticks to buy N
buildings.