bug fixes
The input replaces e + with e on load. I became too hysterical due to input problems, but my spirit recovered.
When loading, worlds, wowCubes, wowTesseracts, wowHypercubes and wowPlatonicCubes are set to 0 if NaN or null.
If you haven't purchased w1x8, Reincarnation Auto Toggle has stopped disabling it by loading. Added w1x8 to the operation judgment of Reincarnation Auto Toggle. This works the same as any other toggle.
Decimal some calculations that are approaching a calculation overflow.
Reduced the amount of challenge complete processing.
Added player.singularityCount to the number of complete Transcension challenges per tick. This is an estimate, so please add an upgrade.
When Auto Challenge Sweep [OFF], Transcension Challenge and Reincarnation Challenge will not automatically exit the challenge if it is completed to the maximum.