Heteract craft keep auto settings on sing
Auto settings are kept on Singularity.
Auto can be set ON/OFF even when craft is not unlocked but it triggers only when unlocked.
Display that a craft is locked in the colored progress bar.
Removed autoHepteractCount global variable because it created unnecessary complexity: it was updated in several parts of the game but used only once -> replaced by function getAutoHepteractCrafts()