w5x8 ant reset bug
reset calls resetAnts and both check for and award the first ant from w5x8. Additionally, as noted by Kimihiro64[1] in Bug reports, the cost does not get updated.

I removed the check in reset, and added the cost increase in resetAnts.

A different fix could be to not increase the cost and have w5x8 award a firstGeneratedAnts instead.
https://discord.com/channels/677271830838640680/1008852065486917765/1008852067344982086