bug fix achievement 75-77
challengeachievementcheck does not get called inside reset('transcension') once we're at cap. if we reach cap inside a reincarnation challenge, the challengeachievementcheck is called with auto=true and does not get to award achivements 75-77.

I moved the challengeachievementcheck in reset('transcension') outside of the block that gets skipped when we're at cap