Penalty Difficulty Indicators
Added a few indicators to the Penalty tab to indicate:
When a new Penalty will be unlocked.
When your effective singularity count will be sharply increased.

These changes are meant to help players more easily understand when difficulty spikes are to be expected so that they can better prepare for them and not get blindsided by them unexpectedly.

New penalty unlocks are currently indicated via the following message replacing the blank space with "???????? ??????? ????? ??? ?????????? ?? ??? (Sing X)" though this can easily be replaced with something like "A new penalty unlocks in Singularity X" if it is so wished.

The penalty difficulty spikes are indicated via an additional line of text that reads "Your penalties will sharply increase in Singularity X", I am not a good programmer so I believe this could be more elegantly implemented but my baseline can be worked off of.